Migrate the parser code to a new crate
