+++ Cargo.toml
@@ -27,6 +27,7 @@ ip_network_table = "0.2.0"
 chrono = { version = "0.4.10", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+rdap_types = { path = "./rdap_types" }
 
 [dev-dependencies]
 criterion = "0.3"
