+++ rdap_types/test_data/ip_network/ip_network_26.json
@@ -1,0 +1,126 @@
+{
+  "objectClassName" : "ip network",
+  "handle" : "XXXX-RIR",
+  "startAddress" : "2001:db8::",
+  "endAddress" : "2001:db8:0:ffff:ffff:ffff:ffff:ffff",
+  "ipVersion" : "v6",
+  "name": "NET-RTR-1",
+  "type" : "DIRECT ALLOCATION",
+  "country" : "AU",
+  "parentHandle" : "YYYY-RIR",
+  "status" : [ "active" ],
+  "remarks" :
+  [
+    {
+      "description" :
+      [
+        "She sells sea shells down by the sea shore.",
+        "Originally written by Terry Sullivan."
+      ]
+    }
+  ],
+"links" :
+[
+{
+"value" : "http://example.net/ip/2001:db8::/48",
+"rel" : "self",
+"href" : "http://example.net/ip/2001:db8::/48",
+"type" : "application/rdap+json"
+},
+{
+"value" : "http://example.net/ip/2001:db8::/48",
+"rel" : "up",
+"href" : "http://example.net/ip/2001:C00::/23",
+"type" : "application/rdap+json"
+}
+],
+"events" :
+[
+{
+"eventAction" : "registration",
+"eventDate" : "1990-12-31T23:59:59Z"
+},
+{
+"eventAction" : "last changed",
+"eventDate" : "1991-12-31T23:59:59Z"
+}
+],
+"entities" :
+[
+{
+"objectClassName" : "entity",
+"handle" : "XXXX",
+"vcardArray":[
+"vcard",
+[
+["version", {}, "text", "4.0"],
+["fn", {}, "text", "Joe User"],
+["kind", {}, "text", "individual"],
+["lang", {
+"pref":"1"
+}, "language-tag", "fr"],
+["lang", {
+"pref":"2"
+}, "language-tag", "en"],
+["org", {
+"type":"work"
+}, "text", "Example"],
+["title", {}, "text", "Research Scientist"],
+["role", {}, "text", "Project Lead"],
+["adr",
+{ "type":"work" },
+"text",
+[
+"",
+"Suite 1234",
+"4321 Rue Somewhere",
+"Quebec",
+"QC",
+"G1V 2M2",
+"Canada"
+]
+],
+["tel",
+{ "type":["work", "voice"], "pref":"1" },
+"uri", "tel:+1-555-555-1234;ext=102"
+],
+["email",
+{ "type":"work" },
+"text", "joe.user@example.com"
+]
+]
+],
+"roles" : [ "registrant" ],
+"remarks" :
+[
+{
+"description" :
+[
+"She sells sea shells down by the sea shore.",
+"Originally written by Terry Sullivan."
+]
+}
+],
+"links" :
+[
+{
+"value" : "http://example.net/entity/xxxx",
+"rel" : "self",
+"href" : "http://example.net/entity/xxxx",
+"type" : "application/rdap+json"
+}
+],
+"events" :
+[
+{
+"eventAction" : "registration",
+"eventDate" : "1990-12-31T23:59:59Z"
+},
+{
+"eventAction" : "last changed",
+"eventDate" : "1991-12-31T23:59:59Z"
+}
+]
+}
+]
+}
